nim game
It is always me who should take first. Think, if it is 4 or multiples of 4, I start first and  I always end up losing the game. For all other cases, I can manage to win since I am talented. For illustration, if there are 5 stones : I will take only one in order to make it 4 and my opponent loses whatever he takes. For 6, I take 2 in order to make it 4 and so on.