Finding whether a number is happy or not
A number is happy if we eventually get the sum of the digits of the  number as 1. The loop of finding the sum of the digits either ends when it becomes 1 or it keeps a cycle. The number is added in a hashset when it is visited for the first time and if it is encountered again, it means it is a cycle so it is not a happy number.